Additional ggplot2 themes using Hind fonts.

This is an add-on pacakge for hrbrthemes pacakge. It provides seven ggplot2 themes based on https://fonts.google.com/specimen/Hind fonts.

Hind Fonts

Hind is an Open Source typeface supporting the Devanagari and Latin scripts. Developed explicitly for use in User Interface design, the Hind font family includes five styles. More information about the font can be found at https://fonts.google.com/specimen/Hind.

One-time setup

Before using this theme you need to call import_hind() which will import the Hind fonts in your extrafont database. Then you will need to install the Hind fonts in the directory mentioned in the output of import_hind() as per your operating system's way of installing fonts.